backend/requirements.txt
# Backend Requirements
fastapi
PyJWT
uvicorn[standard]
pydantic>=2.7.4
python-dotenv
httpx
tenacity

# AI
langchain
langchain-community
langgraph
langchain-google-genai
langchain-openai
tavily-python

# Observability
arize-phoenix
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-langchain

# Database
sqlalchemy>=2.0
asyncpg # Async driver for PostgreSQL, required by SQLAlchemy for async operations
psycopg[binary] # Synchronous driver for PostgreSQL
pgvector # For vector search capabilities in PostgreSQL
greenlet # Required for SQLAlchemy async support

# Cache
redis>=5  